feat: split LivingFreezeEvent & LivingFrozenEvent
