+++ src/main/java/net/neoforged/neoforge/event/entity/living/LivingFreezeEvent.java
@@ -22,23 +22,17 @@ import net.neoforged.neoforge.common.NeoForge;
  */
 public class LivingFreezeEvent extends LivingEvent {
     private boolean isFreezing;
-    private int ticksRequiredToFreeze;
     private float slowAmount;
-    private float damageAmount;
-    private int damageTickRate; // In ticks
 
     public LivingFreezeEvent(LivingEntity entity, boolean isFreezing) {
         super(entity);
         this.isFreezing = isFreezing;
-        this.ticksRequiredToFreeze = entity.getTicksRequiredToFreeze();
         this.slowAmount = -0.05F;
-        this.damageAmount = 1.0F;
-        this.damageTickRate = 40;
     }
 
     /**
      * If the entity is freezing, its freezing counter will be increased. If it's over the
-     * {@link #getTicksRequiredToFreeze()} threshold, the entity will take damage.<br>
+     * {@link LivingEntity#getTicksRequiredToFreeze()} threshold, the entity will take damage.<br>
      * If the entity is not freezing, its freezing counter will be decreased.
      *
      * @return True if the entity is freezing
@@ -56,38 +50,11 @@ public class LivingFreezeEvent extends LivingEvent {
         isFreezing = freezing;
     }
 
-    /**
-     * @return The number of ticks needed to fully freeze (start applying damage) while freezing.
-     */
-    public int getTicksRequiredToFreeze() {
-        return ticksRequiredToFreeze;
-    }
-
-    public void setTicksRequiredToFreeze(int ticksRequiredToFreeze) {
-        this.ticksRequiredToFreeze = ticksRequiredToFreeze;
-    }
-
     public float getSlowAmount() {
         return slowAmount;
     }
 
     public void setSlowAmount(float slowAmount) {
         this.slowAmount = slowAmount;
-    }
-
-    public float getDamageAmount() {
-        return damageAmount;
-    }
-
-    public void setDamageAmount(float damageAmount) {
-        this.damageAmount = damageAmount;
-    }
-
-    public int getDamageTickRate() {
-        return damageTickRate;
-    }
-
-    public void setDamageTickRate(int damageTickRate) {
-        this.damageTickRate = damageTickRate;
     }
 }
